// frontend/src/app/models/user.ts
export class User{
    korisnickoIme: string;
    prezime: string;
    ime: string;
    lozinka: string;
    mejl:string;
    tip: string;
    adresa: string;
    telefon: string;
    licenca: string;
    spec: string;
    ogranak: string;
    status: string;
    slika: string;
    pregledi: Array<string>;
}