// frontend/src/app/models/izvestaji.ts
export class Izvestaj{
    idPregleda: number;
    nazivPregleda: String;
    datum_vreme: Date;
    lekar: string;
    lekar_ime_prezime: string;
    spec: string;
    pacijent: string;
    pacijent_ime_prezime;
    razlogDolaska: string;
    dijagnoza: string;
    terapija: string;
    kontrola: Date;
}